+++ reducers/ImageCacheReducer.js
